Fixed weird CSS on Home Page
Dunno what happened to the CSS since I pushed to main, but it was scuffed, so I fixed it up a bit. I can't currently access the Duel page, so I don't know if that one is messed up too, so for now it's only the Home Page that I updated.